+++ Flipsider/Content/IO/Graphics/Maps/LeavesMap.cs
@@ -10,7 +10,7 @@ namespace Flipsider
 {
     public class LeavesMap : MapPass
     {
-        public override int Priority => 1;
+        public override float Priority => 1;
         protected override Effect? MapEffect => EffectCache.LeavesEffect;
 
         internal override void OnApplyShader()
